Prettified Dao subfolder

+++ contractsSol5/Dao/EpochUtils.sol
@@ -4,17 +4,19 @@ import "../utils/zeppelin/SafeMath.sol";
 
 
 contract EpochUtils {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
-    uint public epochPeriodInSeconds;
-    uint public firstEpochStartTimestamp;
+    uint256 public epochPeriodInSeconds;
+    uint256 public firstEpochStartTimestamp;
 
-    function getCurrentEpochNumber() public view returns(uint) {
+    function getCurrentEpochNumber() public view returns (uint256) {
         return getEpochNumber(now);
     }
 
-    function getEpochNumber(uint timestamp) public view returns(uint) {
-        if (timestamp < firstEpochStartTimestamp || epochPeriodInSeconds == 0) { return 0; }
+    function getEpochNumber(uint256 timestamp) public view returns (uint256) {
+        if (timestamp < firstEpochStartTimestamp || epochPeriodInSeconds == 0) {
+            return 0;
+        }
         // ((timestamp - firstEpochStartTimestamp) / epochPeriodInSeconds) + 1;
         return ((timestamp.sub(firstEpochStartTimestamp)).div(epochPeriodInSeconds)).add(1);
     }
